// backend/src/common/enums.ts
export enum BarrelStatus {
  DRILLING_PENDING = 'DRILLING_PENDING',
  DRILLING_IN_PROGRESS = 'DRILLING_IN_PROGRESS',
  REAMING_PENDING = 'REAMING_PENDING',
  REAMING_IN_PROGRESS = 'REAMING_IN_PROGRESS',
  RIFLING_PENDING = 'RIFLING_PENDING',
  RIFLING_IN_PROGRESS = 'RIFLING_IN_PROGRESS',
  HEAT_TREAT_PENDING = 'HEAT_TREAT_PENDING',
  HEAT_TREAT_IN_PROGRESS = 'HEAT_TREAT_IN_PROGRESS',
  LAPPING_PENDING = 'LAPPING_PENDING',
  LAPPING_IN_PROGRESS = 'LAPPING_IN_PROGRESS',
  HONING_PENDING = 'HONING_PENDING',
  HONING_IN_PROGRESS = 'HONING_IN_PROGRESS',
  CHAMBERING_PENDING = 'CHAMBERING_PENDING',
  CHAMBERING_IN_PROGRESS = 'CHAMBERING_IN_PROGRESS',
  INSPECTION_PENDING = 'INSPECTION_PENDING',
  INSPECTION_IN_PROGRESS = 'INSPECTION_IN_PROGRESS',
  FINISHING_PENDING = 'FINISHING_PENDING',
  FINISHING_IN_PROGRESS = 'FINISHING_IN_PROGRESS',
  FINAL_QC_PENDING = 'FINAL_QC_PENDING',
  FINAL_QC_IN_PROGRESS = 'FINAL_QC_IN_PROGRESS',
  READY_TO_SHIP = 'READY_TO_SHIP',
  REWORK = 'REWORK',
  SCRAP = 'SCRAP',
  HOLD = 'HOLD',
}

export enum ExceptionCode {
  REWORK = 'REWORK',
  SCRAP = 'SCRAP',
  HOLD = 'HOLD',
  QUALITY_ISSUE = 'QUALITY_ISSUE',
  EQUIPMENT_FAILURE = 'EQUIPMENT_FAILURE',
  MATERIAL_DEFECT = 'MATERIAL_DEFECT',
}

export enum StationName {
  DRILLING = 'DRILLING',
  REAMING = 'REAMING',
  RIFLING = 'RIFLING',
  HEAT_TREAT = 'HEAT_TREAT',
  LAPPING = 'LAPPING',
  HONING = 'HONING',
  CHAMBERING = 'CHAMBERING',
  INSPECTION = 'INSPECTION',
  FINISHING = 'FINISHING',
  FINAL_QC = 'FINAL_QC',
}